Decouple storages (#60)
* Add schema manager

* Make console dependency optional in composer

* Actualize schema manager [skip ci]

* Use logger for dumping schema

* Sync CHANGELOG and README with rbac-db

* Update README [skip ci]

* Actualize composer.json

* Add logger for other dbms, uncomment

* Rename

* Rename tests for db schema manager

* Rename and adjust ensure tables multiple test

* Avoid auto generated name for index

* Avoid auto generated name for foreign keys as well

* Avoid auto generated name for foreign keys indexes as well

* Avoid auto generated name for enums as well

* Revert changing enum's name

* Enum: column name = constraint name

* Do not use enum at all because constraint name is hardcoded

* Apply fixes from StyleCI

* Better sync with rbac-db, fix psalm

* Apply fixes from StyleCI

* Revert indexes for fks

* Add sql for MySQL

* Add sql for PostgreSQL [skip ci]

* Add sql for MSSQL [skip ci]

* Add sql for SQLite [skip ci]

* Add sql for migrations [skip ci]

* Disable logger

* Better testing of fks and indexes

* Apply fixes from StyleCI

* Fix locking db in tests

* Reorder protected methods

* Update CHANGELOG.md


* Update examples/Command/RbacCycleInit.php


* Test SQL schema

* Apply fixes from StyleCI

* Reduce copy paste a bit

* Cover the rest of RDBMS

* Debug up queries

* Attempt to fix Windows new lines

* Remove var_dump and extra semicolon

* Try to fix mutation

* Debug

* Revert

* Debug

* Revert

* Update mutation workflow

* Fix duplicate workflow runs

* Test drop schema

* Fix copy paste bug

* Add disconnect

* Test ensureNoTables

* Move disconnect to base testcase

* Debug

* Adjust setup

* Remove debug

* Trigger CI

* Add another disconnect

* Use memory connection

* Empty queries before loop

* Decouple storages in db schema manager and sql schema

* Apply fixes from StyleCI

* Further adjust DbSchemaManager

* Apply fixes from StyleCI

* Revert some of the changes

* Fix Psalm

* More tests for exception

* Apply fixes from StyleCI

* Revert

* Revert revert, remove the rest of unused code

* Fix AssignmentsStorage::renameItem()

* Start testing manager

* Adjust unit tests

* Fix Psalm

* Use common psalm type, add it to parents and children

* Update CI

* Update composer, fix tests, test assignments combination

* Add coupled manager test [skip ci]

* Add manager tests for other RDBMS

* Start to test transaction

* Apply fixes from StyleCI

* Add extension to scrutinizer

* Close connection in tearDown

* Test update permission transaction

* Fix mutant with default value

* Fix commit transaction mutants

* Update docs [skip ci]

* Remove rbac-php dependency [skip ci]

* Actualize migrations

* Improve schema tests for migrations

* Use decorator

* Apply fixes from StyleCI

* Update CHANGELOG

* Sync type hints

* Apply fixes from StyleCI

* Update README [skip ci]

* Sync type hints

* Revert PHPDoc, finalize decorator

* Use short transaction syntax

* Reorder

* Sync with changes in rbac package

* Sync type hinting

* Apply fixes from StyleCI

* Remove runtime [skip ci]

* Change rbac branch to master

* Move common psalm types to rbac package

* Remove duplicated psalm return types

* Partial revert

* Back to master again

* Add info about using manager without decorator [skip ci]

* Revert readme changes [skip ci]

* More detailed changelog [skip ci]

---------

+++ src/DbSchemaManager.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Yiisoft\Rbac\Cycle;
 
 use Cycle\Database\DatabaseInterface;
-use Cycle\Database\ForeignKeyInterface;
 use Cycle\Database\Table;
 use InvalidArgumentException;
 
@@ -16,37 +15,41 @@ use InvalidArgumentException;
 final class DbSchemaManager
 {
     /**
-     * @var string A name of the table for storing RBAC items (roles and permissions).
-     * @psalm-var non-empty-string
+     * @var string|null A name of the table for storing RBAC items (roles and permissions).
+     * @psalm-var ?non-empty-string
      */
-    private string $itemsTable;
+    private ?string $itemsTable;
     /**
-     * @var string A name of the table for storing RBAC assignments.
-     * @psalm-var non-empty-string
+     * @var string|null A name of the table for storing RBAC assignments.
+     * @psalm-var ?non-empty-string
      */
-    private string $assignmentsTable;
+    private ?string $assignmentsTable;
     /**
-     * @var string A name of the table for storing relations between RBAC items.
-     * @psalm-var non-empty-string
+     * @var string|null A name of the table for storing relations between RBAC items.
+     * @psalm-var ?non-empty-string
      */
-    private string $itemsChildrenTable;
+    private ?string $itemsChildrenTable;
 
     /**
-     * @param string $itemsTable A name of the table for storing RBAC items (roles and permissions).
-     * @param string $assignmentsTable A name of the table for storing RBAC assignments.
      * @param DatabaseInterface $database Cycle database instance.
+     * @param string|null $itemsTable A name of the table for storing RBAC items (roles and permissions).
      * @param string|null $itemsChildrenTable A name of the table for storing relations between RBAC items. When set to
      * `null`, it will be automatically generated using {@see $itemsTable}.
+     * @param string|null $assignmentsTable A name of the table for storing RBAC assignments.
      *
      * @throws InvalidArgumentException When a table name is set to the empty string.
      */
     public function __construct(
-        string $itemsTable,
-        string $assignmentsTable,
         private DatabaseInterface $database,
-        string|null $itemsChildrenTable = null,
+        ?string $itemsTable = null,
+        ?string $itemsChildrenTable = null,
+        ?string $assignmentsTable = null,
     ) {
-        $this->initTables($itemsTable, $assignmentsTable, $itemsChildrenTable);
+        $this->initTables(
+            itemsTable: $itemsTable,
+            itemsChildrenTable: $itemsChildrenTable,
+            assignmentsTable: $assignmentsTable,
+        );
     }
 
     /**
@@ -56,6 +59,10 @@ final class DbSchemaManager
      */
     public function createItemsTable(): void
     {
+        if ($this->itemsTable === null || $this->hasTable($this->itemsTable)) {
+            return;
+        }
+
         /** @var Table $table */
         $table = $this->database->table($this->itemsTable);
         $schema = $table->getSchema();
@@ -79,6 +86,14 @@ final class DbSchemaManager
      */
     public function createItemsChildrenTable(): void
     {
+        if (
+            $this->itemsTable ===null ||
+            $this->itemsChildrenTable === null ||
+            $this->hasTable($this->itemsChildrenTable)
+        ) {
+            return;
+        }
+
         /** @var Table $table */
         $table = $this->database->table($this->itemsChildrenTable);
         $schema = $table->getSchema();
@@ -109,6 +124,10 @@ final class DbSchemaManager
      */
     public function createAssignmentsTable(): void
     {
+        if ($this->assignmentsTable === null || $this->hasTable($this->assignmentsTable)) {
+            return;
+        }
+
         /** @var Table $table */
         $table = $this->database->table($this->assignmentsTable);
         $schema = $table->getSchema();
@@ -117,14 +136,6 @@ final class DbSchemaManager
         $schema->string('userId', 128)->nullable(false);
         $schema->setPrimaryKeys(['itemName', 'userId']);
         $schema->integer('createdAt')->nullable(false);
-
-        $schema
-            ->foreignKey(['itemName'])
-            ->references($this->itemsTable, ['name'])
-            ->onUpdate(ForeignKeyInterface::CASCADE)
-            ->onDelete(ForeignKeyInterface::CASCADE)
-            ->setName("fk-$this->assignmentsTable-itemName");
-        $schema->renameIndex(['itemName'], "idx-$this->assignmentsTable-itemName");
 
         $schema->save();
     }
@@ -172,17 +183,9 @@ final class DbSchemaManager
      */
     public function ensureTables(): void
     {
-        if (!$this->hasTable($this->itemsTable)) {
-            $this->createItemsTable();
-        }
-
-        if (!$this->hasTable($this->itemsChildrenTable)) {
-            $this->createItemsChildrenTable();
-        }
-
-        if (!$this->hasTable($this->assignmentsTable)) {
-            $this->createAssignmentsTable();
-        }
+        $this->createItemsTable();
+        $this->createItemsChildrenTable();
+        $this->createAssignmentsTable();
     }
 
     /**
@@ -191,15 +194,15 @@ final class DbSchemaManager
      */
     public function ensureNoTables(): void
     {
-        if ($this->hasTable($this->itemsChildrenTable)) {
+        if ($this->itemsChildrenTable !== null && $this->hasTable($this->itemsChildrenTable)) {
             $this->dropTable($this->itemsChildrenTable);
         }
 
-        if ($this->hasTable($this->assignmentsTable)) {
+        if ($this->assignmentsTable !== null && $this->hasTable($this->assignmentsTable)) {
             $this->dropTable($this->assignmentsTable);
         }
 
-        if ($this->hasTable($this->itemsTable)) {
+        if ($this->itemsTable !== null && $this->hasTable($this->itemsTable)) {
             $this->dropTable($this->itemsTable);
         }
     }
@@ -207,11 +210,11 @@ final class DbSchemaManager
     /**
      * Gets name of the table for storing RBAC items (roles and permissions).
      *
-     * @return string Table name
+     * @return string|null Table name.
      *
      * @see $itemsTable
      */
-    public function getItemsTable(): string
+    public function getItemsTable(): ?string
     {
         return $this->itemsTable;
     }
@@ -219,11 +222,11 @@ final class DbSchemaManager
     /**
      * Gets name of the table for storing RBAC assignments.
      *
-     * @return string Table name.
+     * @return string|null Table name.
      *
      * @see $assignmentsTable
      */
-    public function getAssignmentsTable(): string
+    public function getAssignmentsTable(): ?string
     {
         return $this->assignmentsTable;
     }
@@ -231,11 +234,11 @@ final class DbSchemaManager
     /**
      * Gets name of the table for storing relations between RBAC items.
      *
-     * @return string Table name
+     * @return string|null Table name.
      *
      * @see $itemsChildrenTable
      */
-    public function getItemsChildrenTable(): string
+    public function getItemsChildrenTable(): ?string
     {
         return $this->itemsChildrenTable;
     }
@@ -245,8 +248,12 @@ final class DbSchemaManager
      *
      * @throws InvalidArgumentException When a table name is set to the empty string.
      */
-    private function initTables(string $itemsTable, string $assignmentsTable, string|null $itemsChildrenTable): void
+    private function initTables(?string $itemsTable, ?string $itemsChildrenTable, ?string $assignmentsTable): void
     {
+        if ($itemsTable === null && $assignmentsTable === null) {
+            throw new InvalidArgumentException('At least items table or assignments table name must be set.');
+        }
+
         if ($itemsTable === '') {
             throw new InvalidArgumentException('Items table name can\'t be empty.');
         }
@@ -263,6 +270,8 @@ final class DbSchemaManager
             throw new InvalidArgumentException('Items children table name can\'t be empty.');
         }
 
-        $this->itemsChildrenTable = $itemsChildrenTable ?? $this->itemsTable . '_child';
+        $this->itemsChildrenTable = $itemsTable !== null && $itemsChildrenTable === null
+            ? $itemsTable . '_child'
+            : $itemsChildrenTable;
     }
 }
